backend validation

+++ TaskBoard.WebAPI/Controllers/CardController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using TaskBoard.BLL.Interfaces.Services;
 using TaskBoard.BLL.Models.InputModels;
+using TaskBoard.WebAPI.Validation;
 
 namespace TaskBoard.WebAPI.Controllers;
 
@@ -28,6 +29,13 @@ public class CardController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> Add([FromBody] CardInputModel input)
     {
+        var error = Validator.Card(input);
+
+        if (error.Length > 0)
+        {
+            return BadRequest(error);
+        }
+        
         await _cardService.AddAsync(input);
 
         var cardLists = await _cardListService.GetAsync();
@@ -38,6 +46,13 @@ public class CardController : ControllerBase
     [HttpPatch]
     public async Task<IActionResult> UpdateCard([FromBody] CardInputModel input)
     {
+        var error = Validator.Card(input);
+
+        if (error.Length > 0)
+        {
+            return BadRequest(error);
+        }
+        
         await _cardService.UpdateEntityAsync(input);
 
         var cardLists = await _cardListService.GetAsync();
@@ -48,6 +63,11 @@ public class CardController : ControllerBase
     [HttpPatch("{cardId}/list/{listId}")]
     public async Task<IActionResult> UpdateListInCard(Guid cardId, Guid listId)
     {
+        if (cardId == Guid.Empty || listId == Guid.Empty)
+        {
+            return BadRequest("Card id or List id is empty");
+        }
+        
         await _cardService.UpdateListAsync(cardId, listId);
 
         var cardLists = await _cardListService.GetAsync();
@@ -58,6 +78,11 @@ public class CardController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete([FromRoute] Guid id)
     {
+        if (id == Guid.Empty)
+        {
+            return BadRequest("Card id is empty");
+        }
+        
         await _cardService.DeleteAsync(id);
 
         var cardLists = await _cardListService.GetAsync();
