+++ TaskBoard.WebAPI/Controllers/CardListController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using TaskBoard.BLL.Interfaces.Services;
 using TaskBoard.BLL.Models.InputModels;
+using TaskBoard.WebAPI.Validation;
 
 namespace TaskBoard.WebAPI.Controllers;
 
@@ -34,6 +35,13 @@ public class CardListController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> Add([FromBody] CardListInputModel input)
     {
+        var error = Validator.CardList(input);
+
+        if (error.Length > 0)
+        {
+            return BadRequest(error);
+        }
+        
         await _cardListService.AddAsync(input);
 
         var models = await _cardListService.GetAsync();
@@ -44,6 +52,13 @@ public class CardListController : ControllerBase
     [HttpPut]
     public async Task<IActionResult> Update([FromBody] CardListInputModel input)
     {
+        var error = Validator.CardList(input);
+
+        if (error.Length > 0)
+        {
+            return BadRequest(error);
+        }
+
         await _cardListService.UpdateAsync(input);
 
         var models = await _cardListService.GetAsync();
@@ -54,6 +69,11 @@ public class CardListController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(Guid id)
     {
+        if (id == Guid.Empty)
+        {
+            return BadRequest("CardList id is empty");
+        }
+        
         await _cardListService.DeleteAsync(id);
 
         var models = await _cardListService.GetAsync();
